clients/python/requirements.txt
cryptography>=41.0.0
requests>=2.31.0
PyNaCl>=1.5.0
